requirements.txt
Pillow
mss
opencv-python
numpy
flask
flask_socketio
pywin32
eventlet